+++ tests/Feature/ExampleTest.php
@@ -16,6 +16,6 @@ class ExampleTest extends TestCase
 
         $response->assertStatus(200);
         $response->assertSee('Scholara');
-        $response->assertSee('Librarian');
+        $response->assertSee('Learner');
     }
 }